requirements.txt
Authlib==1.2.1
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==42.0.5
ctransformers==0.2.27
decorator==5.1.1
filelock==3.12.2
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==6.7.0
packaging==23.2
py-cpuinfo==9.0.0
pycparser==2.21
PyYAML==6.0.1
requests==2.28.2
tqdm==4.66.2
typing-extensions==4.7.1
urllib3==1.26.18
validators==0.20.0
weaviate-client==3.15.6
zipp==3.15.0
